Start using regular stdint/stdbool types

+++ Sources/Engine/Base/CRC.cpp
@@ -18,7 +18,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 // Note: this CRC calculation algorithm, although originating from MSDN examples,
 // is in fact identical to the Adler32 used in ZIP's CRC calculation.
 
-ENGINE_API ULONG crc_aulCRCTable[256] = {
+ENGINE_API uint32_t crc_aulCRCTable[256] = {
    0x00000000, 0x77073096, 0xEE0E612C, 0x990951BA, 0x076DC419, 0x706AF48F,
 	 0xE963A535, 0x9E6495A3, 0x0EDB8832, 0x79DCB8A4, 0xE0D5E91E, 0x97D2D988,      
 	 0x09B64C2B, 0x7EB17CBD, 0xE7B82D07, 0x90BF1D91, 0x1DB71064, 0x6AB020F2,
